add support for session password
This is just a basic implementation of a password protected session. I believe we should separate the password from raid settings but that would require a considerable re-write of the logic, so I opted to just use the existing data flow to implement this. If you already have something in the work that is better than this, feel free to close the PR.